added user router and split auth endpoints
