+++ src/routes/users/register.js
@@ -5,7 +5,7 @@ const bcrypt = require('bcrypt')
 const jwt = require('jsonwebtoken');
 
 // Register enpoint
-router.post('/register', async (req, res) => {
+router.post('/', async (req, res) => {
     // Checks whether username or email already exists
     const emailExists = await User.exists( { 'credentials.email': req.body.email });
     if (emailExists) return res.status(409).send({ error: { message: 'Email already in use' }});
@@ -34,30 +34,6 @@ router.post('/register', async (req, res) => {
     res.status(201).send({
         _id: newUser._id,
         username: newUser.credentials.username,
-        token: token
-    });
-});
-
-
-// Login enpoint
-router.post('/login', async (req, res) => {
-    // Check whether username or email exists
-    const user = await User.findOne({ $or: [
-        { 'credentials.email': req.body.email },
-        { 'credentials.username': req.body.username } ]
-    });
-    if (!user) return res.status(404).send({ error: { message: 'Password or username is wrong' }});
-
-    // Compare password to saved hash
-    const correctPassword = await bcrypt.compare(req.body.password, user.credentials.password)
-    if (!correctPassword) return res.status(404).send({ error: { message: 'Password or username is wrong' }});
-
-    // Generate token
-    const token = jwt.sign({ _id: user._id }, process.env.JWT_SECRET);
-
-    res.status(200).send({
-        _id: user._id,
-        username: user.credentials.username,
         token: token
     });
 });
